Added some player characteristics
This push I started creating the framework for our game.

I added the ability to create a player and add stats to him, as well as give him a ClassType

I've also created quite a few blank methods that will slowly be filled in as time goes on.

Created Classes:
1. AnotherEnemy
2. LivingEntity
3. MyPlayer
4. Paladin

Created Interfaces:
1. IAttackBehavior
2. IDialogueBehavior
3. IStats
4. IClassType