Class structure, attacking, and health bars implemented
I finally got the class structure down, and it is working very well!  Enemies and players are easy to create and are being instantiated just how we wanted them to be.

Also fixed the issue with movement so that it is working with the new class structure.

Players are now able to attack with the new class structure.

I have also implemented health bars that drop, with a text value that is clearly visible.

BUG: There is something funky going on with Enemy1 coming up as null.  Looking into it further.