Fixed issue with enemies now spawning properly
The problem was within the game manager prefab - Somehow one of the enemies had been taken out of the array within the prefab, which was calling a null exception error.  Just go into the prefab and put the enemy back in.

Attempting to add another enemy - Skeleton.  Still a work in progress atm.