Fixed bugs, UI stuff, organized, added comments
Fixed a bug where the stats were not being seens correctly.

Changed it so that skills get set to their GUI buttons when the class is instantiated (instead of on the player)

Cleaned up the code, organized some classes, and got rid of some unnecessary methods and classes

Added a bunch of method comments to a large chunk of the project.